requirements.txt
django
django-rest-framework
black
mypy
isort
django-cpf-cnpj
django-phonenumber-field[phonenumbers]
django-environ
django-rest-knox
drf-spectacular